+++ R/utils-randomize-matrix.R
@@ -7,8 +7,8 @@
 #' @param randomize_type How to randomize.
 #'
 #' @return Randomized matrix
-#' @export
 #' @examples
+#' \dontrun{
 #' mat <- matrix(seq_len(9), ncol = 3)
 #' mat
 #'
@@ -17,13 +17,14 @@
 #'
 #' set.seed(42)
 #' randomize_matrix(mat, randomize_type = "cols_independently")
-randomize_matrix <- function(
-    mat,
-    randomize_type = c("rows", "cols_independently")) {
+#' }
+#' @keywords internal
+randomize_matrix <- function(mat,
+                             randomize_type = c("rows", "cols_independently")) {
     randomize_type <- match.arg(randomize_type)
 
     switch(randomize_type,
-        rows = mat[sample(nrow(mat)), ],
+        rows = mat[sample(nrow(mat)), , drop = FALSE],
         cols_independently = apply(mat, 2, sample)
     ) %>%
         `row.names<-`(rownames(mat))